+++ test/test_op1d3t.jl
@@ -1,6 +1,6 @@
 
 using Polynomials4ML, Test
-using Polynomials4ML: evaluate, evaluate_d, evaluate_dd
+using Polynomials4ML: evaluate, evaluate_d, evaluate_d2
 using Polynomials4ML.Testing: println_slim, test_derivatives
 using LinearAlgebra: I
 using QuadGK
