+++ test/test_rtrig.jl
@@ -1,5 +1,5 @@
 using Polynomials4ML, Test
-using Polynomials4ML: evaluate, evaluate_d, evaluate_dd
+using Polynomials4ML: evaluate, evaluate_d, evaluate_d2
 using Polynomials4ML.Testing: println_slim, print_tf, test_derivatives
