+++ test/test_trig.jl
@@ -1,7 +1,7 @@
 
 
 using Polynomials4ML, Test
-using Polynomials4ML: evaluate, evaluate_d, evaluate_dd
+using Polynomials4ML: evaluate, evaluate_d, evaluate_d2
 using Polynomials4ML.Testing: println_slim, print_tf, test_derivatives
